Add min to size interface (#15)
Our use of Shape and Size objects relies heavily on these having a min; each of the size classes has a fairly well-defined min, therefore rather than relying on calling code to check the shape or target shape, it is easier for us to add a min to the Size interface and propagate calls from Shape down to the individual Size objects if necessary (eg, in an 0.5 object).

Also convert FixedSize to a record.